@@ -14,36 +14,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/compiler-cli": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/platform-server": "^5.2.11",
-    "@angular/router": "^5.2.11",
-    "@asymmetrik/angular2-leaflet": "2.1.5",
+    "@angular/animations": "^6.1.10",
+    "@angular/common": "^6.1.10",
+    "@angular/compiler": "^6.1.10",
+    "@angular/compiler-cli": "^6.1.10",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
+    "@angular/platform-browser": "^6.1.10",
+    "@angular/platform-browser-dynamic": "^6.1.10",
+    "@angular/platform-server": "^6.1.10",
+    "@angular/router": "^6.1.10",
+    "@asymmetrik/ngx-leaflet": "3.1.0",
     "@ng-bootstrap/ng-bootstrap": "1.0.0",
     "@types/node": "8.0.14",
     "@types/x2js": "3.1.0",
     "core-js": "2.4.1",
     "font-awesome": "^4.7.0",
-    "jasmine-spec-reporter": "^4.1.0",
+    "leaflet": "^1.7.1",
     "ng-select": "1.0.0-beta.5",
-    "ng2-smart-table": "1.2.2",
-    "rxjs": "^5.5.12",
+    "ng2-smart-table": "1.3.0",
+    "rxjs": "^6.6.3",
+    "rxjs-compat": "^6.0.0-rc.0",
     "ts-helpers": "1.1.1",
     "x2js": "3.1.1",
-    "zone.js": "0.8.17"
+    "zone.js": "0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "^6.2.9",
     "@types/jasmine": "2.5.54",
-    "@types/leaflet": "1.0.66",
-    "codelyzer": "3.1.2",
+    "@types/leaflet": "1.5.19",
+    "codelyzer": "4.5.0",
     "jasmine-core": "2.8.0",
     "jasmine-spec-reporter": "4.1.1",
     "karma": "1.7.0",
@@ -51,11 +53,11 @@
     "karma-cli": "1.0.1",
     "karma-jasmine": "1.0.2",
     "karma-remap-istanbul": "0.6.0",
-    "leaflet": "latest",
     "protractor": "5.1.2",
     "retyped-xml2js-tsd-ambient": "0.0.0-0",
     "ts-node": "3.3.0",
-    "typescript": "2.4.2",
-    "webpack": "3.11.0"
+    "tslint": "^5.5.0",
+    "typescript": "2.9.2",
+    "webpack": "^4.4.1"
   }
 }
