@@ -7,7 +7,7 @@ import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 import { AppComponent } from './app.component';
 import { routing, appRoutingProviders } from './app.routing';
 import { NgbModule, NgbActiveModal, NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
-import { LeafletModule } from "@asymmetrik/angular2-leaflet";
+import { LeafletModule } from "@asymmetrik/ngx-leaflet";
 import { SelectModule } from 'ng-select';
 
 import { HeaderComponent } from './header/header.component'
